requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.46.1
certifi==2025.7.14
charset-normalizer==3.4.2
cut-cross-entropy==25.1.1
datasets==3.6.0
diffusers==0.34.0
dill==0.3.8
docstring_parser==0.16
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
hf-xet==1.1.5
hf_transfer==0.1.9
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.3.1
peft==0.16.0
pillow==11.0.0
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
pyarrow==20.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
sentencepiece==0.2.0
shtab==1.7.2
six==1.17.0
sympy==1.13.1
tokenizers==0.21.2
torch==2.6.0
torchaudio==2.5.1+cu121
torchvision==0.21.0
tqdm==4.67.1
transformers==4.52.4
triton==3.2.0
trl==0.19.1
typeguard==4.4.4
typing_extensions==4.14.1
tyro==0.9.26
tzdata==2025.2
unsloth==2025.7.3
unsloth_zoo==2025.7.4
urllib3==2.5.0
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
